+++ app/src/main/java/domain/Reservation.kt
@@ -1,4 +1,4 @@
-package movie
+package domain
 
 data class Reservation(
     val movie: Movie,
