+++ app/src/test/java/domain/ReservationTest.kt
@@ -1,4 +1,4 @@
-package movie
+package domain
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
